Refactor system
